Fixed build.xml and build system to support optimization without obfuscation and setting additional keep classes.
Unless OBFUSCATION symbol is set, obfuscation will not be activated. Not
even for release builds. Optimizations, however, will be activated for
release builds. Both options will never be activated for debug builds.

In addition it is now possible to set classes for the ProGuard keep
option by adding their full name with package to the @keep array in the
build configuration.